change shareNavcomponent to footer component and edit some css
